merge configurable_dev into configurable

+++ utility/I2CFirmata.h
@@ -18,8 +18,8 @@
 #define I2CFirmata_h
 
 #include <Firmata.h>
-#include <utility/FirmataFeature.h>
-#include <utility/FirmataReporting.h>
+#include "FirmataFeature.h"
+#include "FirmataReporting.h"
 
 // move the following defines to Firmata.h?
 #define I2C_WRITE B00000000
@@ -36,7 +36,7 @@
 /* i2c data */
 struct i2c_device_info {
   byte addr;
-  byte reg;
+  int reg;
   byte bytes;
 };
